bugFix/livedata_emptyList ***  - Rename files.  - UnitTest fixed.
