+++ app/src/main/java/com/alxdev/two/moneychanger/AppApplication.kt
@@ -4,5 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class AppApplication : Application() {
-}
+class AppApplication : Application()
