+++ app/src/main/java/com/alxdev/two/moneychanger/core/dataimp/external/CountryAPIAction.kt
@@ -1,4 +1,4 @@
-package com.alxdev.two.moneychanger.core.data.external
+package com.alxdev.two.moneychanger.core.dataimp.external
 
 import com.alxdev.two.moneychanger.data.remote.currency.ContryDTO
 import ru.gildor.coroutines.retrofit.Result
