+++ app/src/main/java/com/alxdev/two/moneychanger/core/dataimp/local/HistoryDAOAction.kt
@@ -1,4 +1,4 @@
-package com.alxdev.two.moneychanger.core.data.local
+package com.alxdev.two.moneychanger.core.dataimp.local
 
 import androidx.lifecycle.LiveData
 import com.alxdev.two.moneychanger.data.local.entity.History
