+++ build.gradle
@@ -4,10 +4,9 @@ buildscript {
     ext {
         kotlin_version = '1.4.21'
         archLifecycleVersion = '1.1.1'
-        room_version = '2.2.6'
-        coroutine_version = '1.3.8'
         gradleVersion = '4.1.1'
         navigationVersion = '1.0.0'
+        room_version = '2.3.0'
         dataBindingCompilerVersion = gradleVersion // Always need to be the same.
     }
